image-renderer-function/requirements.txt
functions-framework==3.*
Pillow==9.2.0